(chg)-inactmon now uses socket file(AF_UNIX :D) (chg)-removed queueing and client individual threads
Now there is a listen thread and ONE clients thread which send the
message to each connected socket. This fixes the backlog problem with
the multiple queues and simplifies the app. Still doesnt recycle sockets
which are disconnected without trying to operate on them.

Maybe it would also be interesting to be able to use AF_INET sockets if
so chosen.